move `lowering` out from `rust_ir`
