+++ src/bin/chalki.rs
@@ -18,10 +18,10 @@ use std::process::exit;
 use std::sync::Arc;
 
 use chalk::db::ChalkDatabase;
+use chalk::lowering::*;
 use chalk::program_environment::ProgramEnvironment;
 use chalk::query::LoweringDatabase;
 use chalk::rust_ir;
-use chalk::rust_ir::lowering::*;
 use chalk_engine::fallible::NoSolution;
 use chalk_solve::ext::*;
 use chalk_solve::solve::SolverChoice;
